+++ cmd/snap/cmd_services.go
@@ -29,6 +29,7 @@ import (
 	"github.com/snapcore/snapd/client/clientutil"
 	"github.com/snapcore/snapd/i18n"
 	"github.com/snapcore/snapd/snap"
+	"github.com/snapcore/snapd/strutil"
 )
 
 type svcStatus struct {
@@ -93,17 +94,17 @@ func init() {
 		}), argdescs)
 
 	addCommand("start", shortStartHelp, longStartHelp, func() flags.Commander { return &svcStart{} },
-		waitDescs.also(map[string]string{
+		waitDescs.also(userAndScopeDescs).also(map[string]string{
 			// TRANSLATORS: This should not start with a lowercase letter.
 			"enable": i18n.G("As well as starting the service now, arrange for it to be started on boot."),
 		}), argdescs)
 	addCommand("stop", shortStopHelp, longStopHelp, func() flags.Commander { return &svcStop{} },
-		waitDescs.also(map[string]string{
+		waitDescs.also(userAndScopeDescs).also(map[string]string{
 			// TRANSLATORS: This should not start with a lowercase letter.
 			"disable": i18n.G("As well as stopping the service now, arrange for it to no longer be started on boot."),
 		}), argdescs)
 	addCommand("restart", shortRestartHelp, longRestartHelp, func() flags.Commander { return &svcRestart{} },
-		waitDescs.also(map[string]string{
+		waitDescs.also(userAndScopeDescs).also(map[string]string{
 			// TRANSLATORS: This should not start with a lowercase letter.
 			"reload": i18n.G("If the service has a reload command, use it instead of restarting."),
 		}), argdescs)
@@ -184,8 +185,65 @@ func (s *svcLogs) Execute(args []string) error {
 	return nil
 }
 
+type userAndScopeMixin struct {
+	System bool   `long:"system"`
+	User   bool   `long:"user"`
+	Users  string `long:"users"`
+}
+
+var userAndScopeDescs = mixinDescs{
+	// TRANSLATORS: This should not start with a lowercase letter.
+	"system": i18n.G("The operation should only affect system services."),
+	// TRANSLATORS: This should not start with a lowercase letter.
+	"user": i18n.G("The operation should only affect user services for the current user."),
+	// TRANSLATORS: This should not start with a lowercase letter.
+	"users": i18n.G("If provided and set to 'all', the operation should affect services for all users."),
+}
+
+func (um *userAndScopeMixin) validateScopes() error {
+	switch {
+	case um.System && um.User:
+		return fmt.Errorf("--system and --user cannot be used in conjunction with each other")
+	case um.Users != "" && um.User:
+		return fmt.Errorf("--user and --users cannot be used in conjunction with each other")
+	case um.Users != "" && um.Users != "all":
+		return fmt.Errorf("only \"all\" is supported as a value for --users")
+	}
+	return nil
+}
+
+func (um *userAndScopeMixin) serviceScope() client.ScopeSelector {
+	switch {
+	case (um.User || um.Users != "") && !um.System:
+		return client.ScopeSelector([]string{"user"})
+	case !(um.User || um.Users != "") && um.System:
+		return client.ScopeSelector([]string{"system"})
+	}
+	return nil
+}
+
+func (um *userAndScopeMixin) serviceUsers() client.UserSelector {
+	switch {
+	case um.User:
+		return client.UserSelector{
+			Selector: client.UserSelectionSelf,
+		}
+	case um.Users == "all":
+		return client.UserSelector{
+			Selector: client.UserSelectionAll,
+		}
+	}
+	// Currently not reachable as um.Users can only be 'all' for now, but when
+	// we introduce support for lists of usernames, this will be hit.
+	return client.UserSelector{
+		Selector: client.UserSelectionList,
+		Names:    strutil.CommaSeparatedList(um.Users),
+	}
+}
+
 type svcStart struct {
 	waitMixin
+	userAndScopeMixin
 	Positional struct {
 		ServiceNames []serviceName `required:"1"`
 	} `positional-args:"yes" required:"yes"`
@@ -196,8 +254,11 @@ func (s *svcStart) Execute(args []string) error {
 	if len(args) > 0 {
 		return ErrExtraArgs
 	}
+	if err := s.validateScopes(); err != nil {
+		return err
+	}
 	names := svcNames(s.Positional.ServiceNames)
-	changeID, err := s.client.Start(names, client.StartOptions{Enable: s.Enable})
+	changeID, err := s.client.Start(names, s.serviceScope(), s.serviceUsers(), client.StartOptions{Enable: s.Enable})
 	if err != nil {
 		return err
 	}
@@ -215,6 +276,7 @@ func (s *svcStart) Execute(args []string) error {
 
 type svcStop struct {
 	waitMixin
+	userAndScopeMixin
 	Positional struct {
 		ServiceNames []serviceName `required:"1"`
 	} `positional-args:"yes" required:"yes"`
@@ -225,8 +287,11 @@ func (s *svcStop) Execute(args []string) error {
 	if len(args) > 0 {
 		return ErrExtraArgs
 	}
+	if err := s.validateScopes(); err != nil {
+		return err
+	}
 	names := svcNames(s.Positional.ServiceNames)
-	changeID, err := s.client.Stop(names, client.StopOptions{Disable: s.Disable})
+	changeID, err := s.client.Stop(names, s.serviceScope(), s.serviceUsers(), client.StopOptions{Disable: s.Disable})
 	if err != nil {
 		return err
 	}
@@ -237,13 +302,14 @@ func (s *svcStop) Execute(args []string) error {
 		return err
 	}
 
-	fmt.Fprintf(Stdout, i18n.G("Stopped.\n"))
+	fmt.Fprintln(Stdout, i18n.G("Stopped."))
 
 	return nil
 }
 
 type svcRestart struct {
 	waitMixin
+	userAndScopeMixin
 	Positional struct {
 		ServiceNames []serviceName `required:"1"`
 	} `positional-args:"yes" required:"yes"`
@@ -254,8 +320,11 @@ func (s *svcRestart) Execute(args []string) error {
 	if len(args) > 0 {
 		return ErrExtraArgs
 	}
+	if err := s.validateScopes(); err != nil {
+		return err
+	}
 	names := svcNames(s.Positional.ServiceNames)
-	changeID, err := s.client.Restart(names, client.RestartOptions{Reload: s.Reload})
+	changeID, err := s.client.Restart(names, s.serviceScope(), s.serviceUsers(), client.RestartOptions{Reload: s.Reload})
 	if err != nil {
 		return err
 	}
@@ -266,7 +335,7 @@ func (s *svcRestart) Execute(args []string) error {
 		return err
 	}
 
-	fmt.Fprintf(Stdout, i18n.G("Restarted.\n"))
+	fmt.Fprintln(Stdout, i18n.G("Restarted."))
 
 	return nil
 }