+++ overlord/servicestate/servicestate.go
@@ -20,7 +20,6 @@
 package servicestate
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -44,115 +43,11 @@ import (
 	"github.com/snapcore/snapd/wrappers"
 )
 
-type UserSelection int
-
-const (
-	UserSelectionList UserSelection = iota
-	UserSelectionSelf
-	UserSelectionAll
-)
-
-// UserSelector is a support structure for correctly translating a way of
-// representing both a list of user-names, and specific keywords like "self"
-// and "all" for JSON marshalling.
-//
-// When "Selector == UserSelectionList" then Names is used as the data source and
-// the data is treated like a list of strings.
-// When "Selector == UserSelectionSelf|UserSelectionAll", then the data source will
-// be a single string that represent this in the form of "self|all".
-type UserSelector struct {
-	Names    []string
-	Selector UserSelection
-}
-
-// UserList returns a decoded list of users which takes any keyword into account.
-// Takes the current user to be able to handle special keywords like 'user'.
-func (us *UserSelector) UserList(currentUser *user.User) ([]string, error) {
-	switch us.Selector {
-	case UserSelectionList:
-		return us.Names, nil
-	case UserSelectionSelf:
-		if currentUser == nil {
-			return nil, fmt.Errorf(`internal error: for "self" the current user must be provided`)
-		}
-		if currentUser.Uid == "0" {
-			return nil, fmt.Errorf(`cannot use "self" for root user`)
-		}
-		return []string{currentUser.Username}, nil
-	case UserSelectionAll:
-		// Empty list indicates all.
-		return nil, nil
-	}
-	return nil, fmt.Errorf("internal error: unsupported selector %d specified", us.Selector)
-}
-
-func (us UserSelector) MarshalJSON() ([]byte, error) {
-	switch us.Selector {
-	case UserSelectionList:
-		return json.Marshal(us.Names)
-	case UserSelectionSelf:
-		return json.Marshal("self")
-	case UserSelectionAll:
-		return json.Marshal("all")
-	default:
-		return nil, fmt.Errorf("internal error: unsupported selector %d specified", us.Selector)
-	}
-}
-
-func (us *UserSelector) UnmarshalJSON(b []byte) error {
-	// Try treating it as a list of usernames first
-	var users []string
-	if err := json.Unmarshal(b, &users); err == nil {
-		us.Names = users
-		us.Selector = UserSelectionList
-		return nil
-	}
-
-	// Fallback to string, which would indicate a keyword
-	var s string
-	if err := json.Unmarshal(b, &s); err != nil {
-		return fmt.Errorf("cannot unmarshal, expected a string or a list of strings")
-	}
-
-	switch s {
-	case "self":
-		us.Selector = UserSelectionSelf
-	case "all":
-		us.Selector = UserSelectionAll
-	default:
-		return fmt.Errorf(`cannot unmarshal, expected one of: "self", "all"`)
-	}
-	return nil
-}
-
-type ScopeSelector []string
-
-func (ss *ScopeSelector) UnmarshalJSON(b []byte) error {
-	var scopes []string
-	if err := json.Unmarshal(b, &scopes); err != nil {
-		return fmt.Errorf("cannot unmarshal, expected a list of strings")
-	}
-
-	if len(scopes) > 2 {
-		return fmt.Errorf("unexpected number of scopes: %v", scopes)
-	}
-
-	for _, s := range scopes {
-		switch s {
-		case "system", "user":
-		default:
-			return fmt.Errorf(`cannot unmarshal, expected one of: "system", "user"`)
-		}
-	}
-	*ss = scopes
-	return nil
-}
-
 type Instruction struct {
-	Action string        `json:"action"`
-	Names  []string      `json:"names"`
-	Scope  ScopeSelector `json:"scope"`
-	Users  UserSelector  `json:"users"`
+	Action string               `json:"action"`
+	Names  []string             `json:"names"`
+	Scope  client.ScopeSelector `json:"scope"`
+	Users  client.UserSelector  `json:"users"`
 	client.StartOptions
 	client.StopOptions
 	client.RestartOptions
@@ -195,10 +90,10 @@ func (i *Instruction) EnsureDefaultScopeForUser(u *user.User) {
 	if len(i.Scope) == 0 {
 		// If root is making this request, implied scopes are all
 		if u.Uid == "0" {
-			i.Scope = ScopeSelector{"system", "user"}
+			i.Scope = client.ScopeSelector{"system", "user"}
 		} else {
 			// Otherwise imply the service scope only
-			i.Scope = ScopeSelector{"system"}
+			i.Scope = client.ScopeSelector{"system"}
 		}
 	}
 }
@@ -216,7 +111,7 @@ func (i *Instruction) validateScope(u *user.User, apps []*snap.AppInfo) error {
 
 func (i *Instruction) validateUsers(u *user.User, apps []*snap.AppInfo) error {
 	// Perform some additional user checks
-	if i.Users.Selector == UserSelectionList && len(i.Users.Names) == 0 {
+	if i.Users.Selector == client.UserSelectionList && len(i.Users.Names) == 0 {
 		// It is an error for a non-root to not specify any users if we are targeting
 		// user daemons
 		if u.Uid != "0" && i.hasUserService(apps) {